requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0

jinja2==3.1.3

sqlalchemy==2.0.25
pymysql==1.1.0

pandas==2.2.0
numpy==1.26.3


python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4

reportlab==4.2.2
